twitter/functions/twitter_ingest_get_tweets/requirements.txt
google-cloud-secret-manager==2.2.0
google-cloud-firestore==2.0.2
google-cloud-pubsub==1.5.0
elasticsearch==7.13.4
